+++ client/components/InterestHub/InterestHub.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import styled from 'styled-components';
 import { PageTitle } from '../CommonStyles.js';
-import Forum from './Forum/Forum.jsx';
+import Forum from '../Forum/Forum.jsx';
 
 const InterestHubContainer = styled.div`
   display: flex;
